chore: ordenar archivos de configuración para logger y kafka para reducir complejidad
