--- crawlers/base/__init__.py
@@ -1,15 +1,0 @@
-"""
-Essential utilities for the News Monitor.
-"""
-
-from .utils import (
-    setup_logging,
-    get_config,
-    normalize_url
-)
-
-__all__ = [
-    'setup_logging',
-    'get_config',
-    'normalize_url'
-]
