+++ crawlers/base/utils.py
@@ -3,35 +3,8 @@ Essential utility functions for the News Monitor.
 """
 
 import os
-import logging
-import structlog
 from urllib.parse import urljoin, urlparse, urlunparse
 from typing import Dict, Any, Optional
-
-
-def setup_logging(level: str = "INFO", service_name: str = "news-monitor") -> None:
-    """Setup simple structured logging."""
-    
-    # Configure structlog for simple, readable output
-    structlog.configure(
-        processors=[
-            structlog.stdlib.filter_by_level,
-            structlog.stdlib.add_log_level,
-            structlog.processors.TimeStamper(fmt="%Y-%m-%d %H:%M:%S"),
-            structlog.dev.ConsoleRenderer()
-        ],
-        context_class=dict,
-        logger_factory=structlog.stdlib.LoggerFactory(),
-        wrapper_class=structlog.stdlib.BoundLogger,
-        cache_logger_on_first_use=True,
-    )
-    
-    # Configure standard logging
-    logging.basicConfig(
-        format="%(message)s",
-        level=getattr(logging, level.upper()),
-    )
-
 
 def get_config() -> Dict[str, Any]:
     """Load configuration from environment variables."""
